pyproject.toml
[project]
name = "rag-search-engine-optimized"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=0.1.0",
    "nltk==3.9.1",
    "numpy>=2.3.5",
    "pillow>=12.1.0",
    "sentence-transformers>=5.2.0",
]
